+++ cmd/ngcore/defaults.go
@@ -5,9 +5,13 @@ const (
 	usage       = "Brand-new golang daemon implement of Ngin Network Node"
 	description = "NGIN is a radically updating brand-new blockchain network, " +
 		"which is not a fork of ethereum or any other chain."
-	version = "v0.0.16"
+	version = "v0.0.17"
 
 	// flag values
-	defaultTCPP2PPort = 52520
-	defaultAPIPort    = 52521
+	defaultTCPP2PPort   = 52520
+	defaultRPCPort      = 52521
+	defaultLogLevel     = "WARN"
+	defaultLogFile      = "ngcore.log"
+	dafaultMiningThread = -1
+	defaultDBFolder     = ".ngdb"
 )