+++ consensus/sync_remote_record.go
@@ -30,6 +30,7 @@ func (r *remoteRecord) shouldSync() bool {
 // RULE: when forking?
 // Situation #1: remote height is higher than local, AND checkpoint is on higher level
 // Situation #2: remote height is higher than local, AND checkpoint is on same level, AND remote checkpoint takes more rank (with more ActualDiff)
+// TODO: add a cap for forking
 func (r *remoteRecord) shouldFork() bool {
 	cp := storage.GetChain().GetLatestCheckpoint()
 	cpHash := cp.Hash()