+++ keytools/tools.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"path/filepath"
 
-	logging "github.com/ipfs/go-log/v2"
 	"github.com/mr-tron/base58"
 
 	"github.com/ngchain/secp256k1"
@@ -17,11 +16,20 @@ import (
 	"github.com/ngchain/ngcore/utils"
 )
 
-var log = logging.Logger("key")
-
 // ReadLocalKey will read the local AES-256-GCM encrypted secp256k1 key file to load an ecdsa private key.
 func ReadLocalKey(filename string, password string) *secp256k1.PrivateKey {
 	var key *secp256k1.PrivateKey
+
+	if filename == "" {
+		path := GetDefaultFolder()
+		if _, err := os.Stat(path); os.IsNotExist(err) {
+			err := os.Mkdir(path, os.ModePerm)
+			if err != nil {
+				panic(err)
+			}
+		}
+		filename = GetDefaultFile()
+	}
 
 	if _, err := os.Stat(filename); err != nil {
 		key = CreateLocalKey(filename, password)
@@ -30,7 +38,7 @@ func ReadLocalKey(filename string, password string) *secp256k1.PrivateKey {
 
 		raw, err = ioutil.ReadFile(filepath.Clean(filename))
 		if err != nil {
-			log.Panic(err)
+			panic(err)
 		}
 
 		rawPK := utils.AES256GCMDecrypt(raw, []byte(password))
@@ -40,24 +48,42 @@ func ReadLocalKey(filename string, password string) *secp256k1.PrivateKey {
 	return key
 }
 
-// CreateLocalKey will create a keyfile named *filename* and encrypted with *password* in aes-256-gcm.
-func CreateLocalKey(filename string, password string) *secp256k1.PrivateKey {
+// NewLocalKey will create a privateKey only
+func NewLocalKey() *secp256k1.PrivateKey {
 	key, err := secp256k1.GeneratePrivateKey()
 	if err != nil {
-		log.Panic(err)
+		panic(err)
+	}
+
+	return key
+}
+
+// CreateLocalKey will create a keyfile named *filename* and encrypted with *password* in aes-256-gcm.
+func CreateLocalKey(filename, password string) *secp256k1.PrivateKey {
+	key := NewLocalKey()
+
+	if filename == "" {
+		path := GetDefaultFolder()
+		if _, err := os.Stat(path); os.IsNotExist(err) {
+			err := os.Mkdir(path, os.ModePerm)
+			if err != nil {
+				panic(err)
+			}
+		}
+		filename = GetDefaultFile()
 	}
 
 	// save key to ngcore.key file
 	file, err := os.Create(filename)
 	if err != nil {
-		log.Panic(err)
+		panic(err)
 	}
 
 	encrypted := utils.AES256GCMEncrypt(key.D.Bytes(), []byte(password))
 
 	_, err = file.Write(encrypted)
 	if err != nil {
-		log.Panic(err)
+		panic(err)
 	}
 
 	_ = file.Close()
@@ -65,14 +91,46 @@ func CreateLocalKey(filename string, password string) *secp256k1.PrivateKey {
 	return key
 }
 
-// PrintAddress will print the privateKey's **address** to the console.
-func PrintAddress(privateKey *secp256k1.PrivateKey) {
-	address := ngtypes.NewAddress(privateKey)
-	log.Warnf("Address is: %s\n", base58.FastBase58Encoding(address))
+// RecoverLocalKey will recover a keyfile named *filename* with the password from the privateKey string.
+func RecoverLocalKey(filename, password, privateKey string) *secp256k1.PrivateKey {
+	bKey, err := base58.FastBase58Decoding(privateKey)
+	if err != nil {
+		panic(err)
+	}
+
+	key := secp256k1.NewPrivateKey(new(big.Int).SetBytes(bKey))
+
+	if filename == "" {
+		path := GetDefaultFolder()
+		if _, err := os.Stat(path); os.IsNotExist(err) {
+			err := os.Mkdir(path, os.ModePerm)
+			if err != nil {
+				panic(err)
+			}
+		}
+		filename = GetDefaultFile()
+	}
+
+	// save key to ngcore.key file
+	file, err := os.Create(filename)
+	if err != nil {
+		panic(err)
+	}
+
+	encrypted := utils.AES256GCMEncrypt(key.D.Bytes(), []byte(password))
+
+	_, err = file.Write(encrypted)
+	if err != nil {
+		panic(err)
+	}
+
+	_ = file.Close()
+
+	return key
 }
 
-// PrintKeyPair will print the **privateKey and its publicKey** to the console.
-func PrintKeyPair(privateKey *secp256k1.PrivateKey) {
+// PrintKeysAndAddress will print the **privateKey and its publicKey** to the console.
+func PrintKeysAndAddress(privateKey *secp256k1.PrivateKey) {
 	rawPrivateKey := privateKey.Serialize() // its D
 	fmt.Println("Private Key: ", base58.FastBase58Encoding(rawPrivateKey))